requirements_min.txt
boto3==1.37.23
botocore==1.37.23
numpy==1.26.4
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.3
fastapi==0.115.8
fastapi-cli==0.0.7
ffmpeg-python==0.2.0
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
yt-dlp==2023.11.14
ultralytics>=8.0.0
uvicorn
pillow
openai
youtube-transcript-api
sqlalchemy